requirements.txt
Pillow=8.3.1
numpy
pynput